requirements.txt
absl-py==0.11.0
apache-beam==2.25.0
appnope==0.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrdict==2.0.1
attrs==20.2.0
avro-python3==1.9.2.1
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.16.16
botocore==1.19.16
bqplot==0.12.18
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
cloudpickle==1.6.0
colour==0.1.5
countrycode==0.2
crcmod==1.7
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docopt==0.6.2
earthengine-api==0.1.240
entrypoints==0.3
fastavro==1.1.0
fasteners==0.15
ffmpeg-python==0.2.0
Fiona==1.8.17
folium==0.11.0
future==0.18.2
gast==0.3.3
geeadd==0.5.2
geemap==0.8.1
geocoder==1.38.1
geographiclib==1.50
geojson==2.5.0
geopandas==0.8.1
geopy==2.0.0
gmaps==0.9.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-apitools==0.5.31
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.3.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
httplib2shim==0.0.3
idna==2.10
importlib-metadata==2.0.0
ipyfilechooser==0.4.0
ipykernel==5.3.4
ipyleaflet==0.13.3
ipynb-py-convert==0.4.6
ipython==7.16.1
ipython-genutils==0.2.0
ipytree==0.1.8
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==0.1.1
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
langid==1.1.6
logzero==1.6.2
loky==2.9.0
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mock==2.0.0
monotonic==1.5
mss==6.1.0
munch==2.5.0
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.8
nest-asyncio==1.4.2
nltk==3.5
notebook==6.1.5
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
overpy==0.4
packaging==20.4
pageviewapi==0.4.0
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip-autoremove==0.9.1
plotly==4.12.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pydot==1.4.1
Pygments==2.7.2
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==3.0.0.post1
pyrsistent==0.17.3
pyshp==2.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytrends==4.7.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
ratelim==0.1.6
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
reverse-geocode==1.4.1
reverse-geocoder==1.5.1
rsa==4.6
Rtree==0.9.4
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.51.0
traitlets==4.3.3
traittypes==0.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
utm==0.6.0
voila==0.1.23
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
